Check achievements on suggest and poll resolution
